fix: unify init command with new feat
