+++ libraries/module/index.ts
@@ -28,5 +28,6 @@ export * from "./lib/extension/Tool";
 
 export * from "./lib/util/Configuration";
 export * from "./lib/util/diagnostics";
+export * from "./lib/util/extractArgumentValues";
 
 export * from "./lib/ModuleManager";
