+++ package.json
@@ -58,7 +58,7 @@
     "@types/seedrandom": "^3.0.5",
     "@types/sinon": "10.0.13",
     "@types/ws": "^8.5.4",
-    "@types/yargs": "17.0.20",
+    "@types/yargs": "17.0.24",
     "@typescript-eslint/eslint-plugin": "5.45.1",
     "@typescript-eslint/parser": "5.45.1",
     "chai": "4.3.7",
